+++ src/lib/Helper/Words/LocalWordsHelper.php
@@ -49,18 +49,11 @@ class LocalWordsHelper extends AbstractWordsHelper {
         $length = $strength == 1 ? 2:$strength;
         $file   = $this->getWordsFile();
 
-        $retires = 0;
-        while($retires < 24) {
+        for($i = 0; $i < 24; $i++) {
             $result = [];
             exec("shuf -n {$length} {$file}", $result, $code);
 
-            if($code == 0) {
-                $max = max(array_map('strlen', $result));
-                $min = min(array_map('strlen', $result));
-
-                if($min > 3 && $max < 12) return $result;
-            }
-            $retires++;
+            if($code == 0 && $this->isWordsArrayValid($result)) return $result;
         }
 
         return [];