+++ src/lib/Services/HelperService.php
@@ -22,6 +22,7 @@ use OCA\Passwords\Helper\Image\ImagickHelper;
 use OCA\Passwords\Helper\Preview\AbstractPreviewHelper;
 use OCA\Passwords\Helper\Preview\DefaultPreviewHelper;
 use OCA\Passwords\Helper\Preview\PageresCliHelper;
+use OCA\Passwords\Helper\Preview\WebshotHelper;
 use OCA\Passwords\Helper\Preview\ScreenShotApiHelper;
 use OCA\Passwords\Helper\Preview\ScreenShotMachineHelper;
 use OCA\Passwords\Helper\SecurityCheck\AbstractSecurityCheckHelper;
@@ -44,6 +45,7 @@ class HelperService {
 
     const PREVIEW_SCREEN_SHOT_MACHINE = 'ssm';
     const PREVIEW_SCREEN_SHOT_API     = 'ssa';
+    const PREVIEW_WEBSHOT             = 'ws';
     const PREVIEW_PAGERES             = 'pageres';
     const PREVIEW_DEFAULT             = 'default';
 
@@ -89,6 +91,7 @@ class HelperService {
 
     /**
      * @return AbstractImageHelper
+     * @throws \OCP\AppFramework\QueryException
      */
     public function getImageHelper(): AbstractImageHelper {
         $service = $this->config->getAppValue('service/images', self::IMAGES_IMAGICK);
@@ -102,6 +105,7 @@ class HelperService {
 
     /**
      * @return AbstractPreviewHelper
+     * @throws \OCP\AppFramework\QueryException
      */
     public function getWebsitePreviewHelper(): AbstractPreviewHelper {
         $service = $this->config->getAppValue('service/preview', self::PREVIEW_DEFAULT);
@@ -109,6 +113,8 @@ class HelperService {
         switch($service) {
             case self::PREVIEW_PAGERES:
                 return $this->container->query(PageresCliHelper::class);
+            case self::PREVIEW_WEBSHOT:
+                return $this->container->query(WebshotHelper::class);
             case self::PREVIEW_SCREEN_SHOT_API:
                 return $this->container->query(ScreenShotApiHelper::class);
             case self::PREVIEW_SCREEN_SHOT_MACHINE:
@@ -122,6 +128,7 @@ class HelperService {
 
     /**
      * @return AbstractFaviconHelper
+     * @throws \OCP\AppFramework\QueryException
      */
     public function getFaviconHelper(): AbstractFaviconHelper {
         $service = $this->config->getAppValue('service/favicon', self::FAVICON_DEFAULT);
@@ -146,9 +153,10 @@ class HelperService {
 
     /**
      * @return AbstractWordsHelper
+     * @throws \OCP\AppFramework\QueryException
      */
     public function getWordsHelper(): AbstractWordsHelper {
-        $service = $this->config->getAppValue('service/words', self::WORDS_RANDOM);
+        $service = $this->config->getAppValue('service/words', null);
 
         switch($service) {
             case self::WORDS_LOCAL:
@@ -159,11 +167,14 @@ class HelperService {
                 return $this->container->query(RandomCharactersHelper::class);
         }
 
+        if(is_file(LocalWordsHelper::WORDS_DEFAULT)) return $this->container->query(LocalWordsHelper::class);
+
         return $this->container->query(RandomCharactersHelper::class);
     }
 
     /**
      * @return AbstractSecurityCheckHelper
+     * @throws \OCP\AppFramework\QueryException
      */
     public function getSecurityHelper(): AbstractSecurityCheckHelper {
         $service = $this->config->getAppValue('service/security', self::SECURITY_HIBP);
@@ -184,6 +195,7 @@ class HelperService {
 
     /**
      * @return DefaultFaviconHelper
+     * @throws \OCP\AppFramework\QueryException
      */
     public function getDefaultFaviconHelper(): DefaultFaviconHelper {
         return $this->container->query(DefaultFaviconHelper::class);