+++ src/lib/Settings/AdminSettings.php
@@ -11,8 +11,10 @@ use Gmagick;
 use Imagick;
 use OCA\Passwords\AppInfo\Application;
 use OCA\Passwords\Helper\Favicon\BestIconHelper;
+use OCA\Passwords\Helper\Preview\WebshotHelper;
 use OCA\Passwords\Helper\Preview\ScreenShotApiHelper;
 use OCA\Passwords\Helper\Preview\ScreenShotMachineHelper;
+use OCA\Passwords\Helper\Words\LocalWordsHelper;
 use OCA\Passwords\Services\ConfigurationService;
 use OCA\Passwords\Services\FileCacheService;
 use OCA\Passwords\Services\HelperService;
@@ -132,7 +134,8 @@ class AdminSettings implements ISettings {
      * @return array
      */
     protected function getWordsServices(): array {
-        $current = $this->config->getAppValue('service/words', HelperService::WORDS_RANDOM);
+        $current = is_file(LocalWordsHelper::WORDS_DEFAULT) ? HelperService::WORDS_LOCAL:HelperService::WORDS_RANDOM;
+        $current = $this->config->getAppValue('service/words', $current);
 
         return [
             [
@@ -240,6 +243,15 @@ class AdminSettings implements ISettings {
                 'api'     => null
             ],
             [
+                'id'      => HelperService::PREVIEW_WEBSHOT,
+                'label'   => 'Passwords Webshot',
+                'current' => $current === HelperService::PREVIEW_WEBSHOT,
+                'api'     => [
+                    'key'   => WebshotHelper::WEBSHOT_CONFIG_KEY,
+                    'value' => $this->config->getAppValue(WebshotHelper::WEBSHOT_CONFIG_KEY)
+                ]
+            ],
+            [
                 'id'      => HelperService::PREVIEW_SCREEN_SHOT_API,
                 'label'   => 'screenshotapi.io',
                 'current' => $current === HelperService::PREVIEW_SCREEN_SHOT_API,
@@ -334,7 +346,7 @@ class AdminSettings implements ISettings {
         $cronType  = \OC::$server->getConfig()->getAppValue('core', 'backgroundjobs_mode', 'ajax');
 
         return [
-            'cron'   => $cronType === 'ajax',
+            'cron'   => $cronType,
             'https'  => \OC::$server->getRequest()->getHttpProtocol() === 'https',
             'php'    => [
                 'warn'    => false,